requirements.txt
# Core ML & CV
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Depth estimation
timm>=0.9.0

# Tracking
supervision>=0.16.0
filterpy>=1.4.5
lap>=0.4.0

# Server & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# CLI & Utils
typer>=0.9.0
rich>=13.7.0
pydantic>=2.0.0
pyyaml>=6.0
orjson>=3.9.0

# Database & Auth
supabase>=2.0.0
python-dotenv>=1.0.0

# Code quality (dev)
black>=23.0.0
ruff>=0.1.0
pytest>=7.4.0
